backend/pyproject.toml
[project]
name = "socitwin-backend"
version = "0.1.0"
description = "Socitwin Backend API"
requires-python = ">=3.11,<3.12"
dependencies = [
    "apify-client>=1.8.0",
    "fastapi>=0.135.3",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.44.0",
    "camel-oasis==0.2.5",
    "chromadb>=1.0.0",
    "openai>=1.0.0",
    "aiofiles>=24.0.0",
    "matplotlib>=3.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "httpx>=0.27.0",
    "ruff>=0.1.0",
    "pyright>=1.1.300",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.ruff.lint.per-file-ignores]
"app/models/agent_monitor.py" = ["N815"]
"tests/e2e/e2e_simulation_test.py" = ["E402"]

[tool.pyright]
include = ["app"]
exclude = ["**/node_modules", "**/__pycache__", "**/.venv"]
typeCheckingMode = "standard"
pythonVersion = "3.11"
stubPath = ""
ignore = ["**/oasis*"]
